==> Code/00_Production_Sequence.ipynb <==
<h2>Production Sequence

<h4>This notebook includes the production sequence of AI usage and acceptance rate data analysis project

<h5>The notebooks related to this project are:

01_Data_Processing<br>
02_Data_Wrangling<br>
03_Data_Visualization<br>
04_Model_Analysis<br>

The rest of this notebook will document the purpose of each of above notebooks

<h4><b>01_Data_Processing

This notebook cleans and processes original survey and factset data.<br>

It deals with data types and null values, while also creating a common sector column for the two datasets to be joined<br>

The final output is two tables:<br>

AI_usage_wth_fin_data: this table contains average AI usage rate (i.e., did this business use Artificial Intelligence (AI) in producing goods or services? ), as well as financial data (i.e., Market Value, Sales, etc.)<br>

AI_acceptance_wth_fin_data: this table contains average AI acceptance rate (i.e., will this business use Artificial Intelligence (AI) in the next 6 months? ), as well as financial data (i.e., Market Value, Sales, etc.)

<h4><b>02_Data_Wrangling

This notebook cleans the data, since the financial data is difficult to normalize, we only round the digits, as well as processed with one missing value for utility sector as none of the firms diclosed their R&D expense.<br>
The average of R&D expense for the rest of the sectors are used for that missing value.

The final output are the following tables:<br>

02_AI_usage_upd: with cleaned data<br>
02_AI_acceptance_upd: with cleaned data<br>

<h4><b>03_Data_Visualization

This notebook generates initial descriptive analysis/visualization result of AI usage and accepctance rate across different sectors, as well as financial data

This notebook contains the following graphs:

Line chart: Overall AI Usage and Acceptance Rate across survey period<br>
Bar chart: AI usage & acceptance by sector<br>
Line & Bar chart: Market Value, EBITDA Margin, and Sales Growth by Sector<br>
Line & Bar chart: ROE, P/E Ratio, and R&D Intensity by Sector<br>
Bubble Chart: Usage and Acceptance Rates VS market value<br>
Bubble Chart: Usage and Acceptance Rates VS growth

This notebook does not generate table(data) output

<h4><b>04_Model_Analysis

This notebook intends to dive deeper into what factor impacts bussiness's decision of using AI<br>

It employs a linear model, analyzes the model fit, and further applies principal component regression<br>

Linear model: linear regression of average AI usage rate on the following financial factors:<br>
avg_Market_Value: average market value<br>
avg_1_Year_Growth_Total_Sales: average growth<br>
avg_FE_Val_Pe_Mean_FY1_Roll: average price-to-equity multiple <br>
avg_Resrch_and_Develop_Expense: average R&D expense

Principal Components Analysis:<br>
Combines the above four dependent variables, which gives an "index" of financial strength<br>
Applies Leave-One-Out regression 

Model analysis result:<br>
The coefficient of financial strength is statistically significant.

However, the authors acknowledge the limitation of this application, as data we obtained is limited<br>

Acknowledgement of External Assistance: <br>
01_Data_Processing and 02_Data_Wrangling are completely written by the authors.<br>

For 03_Data_Visualization, we used ChatGPT to help modify codes for visualization, which gives us intended results (such as line colors, text size, etc.) <br>

For 04_Model_Analysis, we also used ChatGPT to modify codes in principal component analysis, while our original idea is to measure how financial strength influences AI uasge, ChatGPT suggested using Leave-One-Out regression.
