==> Code/01_Data_Processing.ipynb <==
<h2>Data Preprocessing

# Initializing spark session 
from pyspark.sql import SparkSession 
spark = SparkSession.builder.appName("Test").getOrCreate() 

# Import relevant packages 
import pandas as pd
from pyspark.sql import functions as F
from pyspark.sql.functions import * 
from pyspark.sql import DataFrame 
from pyspark.sql.types import StructType, StructField, StringType, DoubleType

# define data path
data_path = "/Users/lanzia/Desktop/Study/FS 2025/01.1st Semester/Introduction to Data in Business Analytics/Project/Data/"

# define the columns needed to be cleaned later
ranges = [(202319, 202326), (202401, 202426), (202501, 202518)]
numeric_fields = [StructField(str(i), DoubleType(), True) 
                  for start, end in ranges for i in range(start, end + 1)
                  ]
cols_to_clean = [str(i) for start, end in ranges for i in range(start, end + 1)]

<h4>Step 1: clean survey data

**Load the tables**

# survery data
btos_survey = pd.read_excel(data_path + "Sector.xlsx",  sheet_name="Response Estimates", engine="openpyxl")
btos_survey = spark.createDataFrame(btos_survey)
btos_survey.show(5, truncate=False)

# Check the Schema
btos_survey.printSchema

# filter the table to keep the rows related to AI usage question
btos_survey_filtered = (btos_survey
                   .filter((col('Question ID')==7) | (col('Question ID')==24))
                   .filter((col('Sector')!='XX'))
                   )

btos_survey_filtered.show(5, truncate=False)

# clean the columns
# for value with 'S', it's equivalent of nulls; convert the rest of percentage to float value
for c in cols_to_clean:
    if c in btos_survey_filtered.columns:
        btos_survey_filtered = (btos_survey_filtered
                                .withColumn(c, F.when(F.col(c).isin("S"), None)
                                            .otherwise(F.regexp_replace(F.trim(F.col(c)), "%", "").cast("double"))
                                            )
                                )

# sample print
btos_survey_filtered.show(10)

# map sector ID to names that match with S&P factset data
btos_survey_fnl = (btos_survey_filtered
                   .withColumnRenamed('Sector','subsector')
                   .withColumn('subsector_name', when(col('subsector')==11,  lit("Agriculture, Forestry, Fishing and Hunting"))
                                              .when(col('subsector')==21, lit("Mining, Quarrying, and Oil and Gas Extraction"))
                                              .when(col('subsector')==22, lit("Utilities"))
                                              .when(col('subsector')==23, lit("Construction"))
                                              .when(col('subsector')==31, lit("Manufacturing"))
                                              .when(col('subsector')==32, lit("Manufacturing"))
                                              .when(col('subsector')==33, lit("Manufacturing"))
                                              .when(col('subsector')==42, lit("Wholesale Trade"))
                                              .when(col('subsector')==44, lit("Retail Trade"))
                                              .when(col('subsector')==45, lit("Retail Trade"))
                                              .when(col('subsector')==48, lit("Transportation and Warehousing"))
                                              .when(col('subsector')==49, lit("Transportation and Warehousing"))
                                              .when(col('subsector')==51, lit("Information"))
                                              .when(col('subsector')==52, lit("Finance and Insurance"))
                                              .when(col('subsector')==53, lit("Real Estate and Rental and Leasing"))
                                              .when(col('subsector')==54, lit("Professional, Scientific, and Technical Services"))
                                              .when(col('subsector')==55, lit("Management of Companies and Enterprises"))
                                              .when(col('subsector')==56, lit("Administrative and Support and Waste Management and Remediation Services"))
                                              .when(col('subsector')==61, lit("Educational Services"))
                                              .when(col('subsector')==62, lit("Health Care and Social Assistance"))
                                              .when(col('subsector')==71, lit("Arts, Entertainment, and Recreation"))
                                              .when(col('subsector')==72, lit("Accommodation and Food Services"))
                                              .when(col('subsector')==81, lit("Other Services (except Public Administration)"))
                                              .when(col('subsector')==92, lit("Public Administration"))
                                              .otherwise(lit(None))
                                )
                    .withColumn('Sector', when(col('subsector_name')=="Agriculture, Forestry, Fishing and Hunting", lit("Consumer Non-Cyclicals"))
                                                                .when(col('subsector_name')=="Mining, Quarrying, and Oil and Gas Extraction", lit("Energy"))
                                                                .when(col('subsector_name')=="Utilities", lit("Utilities"))
                                                                .when(col('subsector_name')=="Construction", lit("Industrials"))
                                                                .when(col('subsector_name')=="Manufacturing", lit("Industrials"))
                                                                .when(col('subsector_name')=="Wholesale Trade", lit("Industrials"))
                                                                .when(col('subsector_name')=="Retail Trade", lit("Consumer Cyclicals"))
                                                                .when(col('subsector_name')=="Transportation and Warehousing", lit("Industrials"))
                                                                .when(col('subsector_name')=="Information", lit("Technology"))
                                                                .when(col('subsector_name')=="Finance and Insurance", lit("Finance"))
                                                                .when(col('subsector_name')=="Real Estate and Rental and Leasing", lit("Finance"))
                                                                .when(col('subsector_name')=="Professional, Scientific, and Technical Services", lit("Business Services"))
                                                                .when(col('subsector_name')=="Management of Companies and Enterprises", lit("Business Services"))
                                                                .when(col('subsector_name')=="Administrative and Support and Waste Management and Remediation Services", lit("Business Services"))
                                                                .when(col('subsector_name')=="Educational Services", lit("Consumer Services"))
                                                                .when(col('subsector_name')=="Health Care and Social Assistance", lit("Healthcare"))
                                                                .when(col('subsector_name')=="Arts, Entertainment, and Recreation", lit("Consumer Services"))
                                                                .when(col('subsector_name')=="Accommodation and Food Services", lit("Consumer Services"))
                                                                .when(col('subsector_name')=="Other Services (except Public Administration)", lit("Consumer Services"))
                                                                .when(col('subsector_name')=="Public Administration", lit(None))
                                                                .otherwise(lit(None)))
                    .select('Sector','subsector','subsector_name','Question ID','Answer', *cols_to_clean)
                    )

# sample print
btos_survey_fnl.show(10)

# check the map
btos_survey_fnl.groupBy('Sector','subsector','subsector_name').count().sort('subsector').show(truncate=False)

# calculate average AI acceptance rate across the same sector
btos_survey_fnlfnl = (btos_survey_fnl
                      .filter(col('Answer')=="Yes")
                      .groupBy('Sector','Question ID')
                      .agg(*[F.avg(c).alias(f"avg_{c.replace(' ', '_')}") for c in cols_to_clean])
                      .sort('Sector','Question ID')
                      )

# sample print
btos_survey_fnlfnl.show()

# check schema
btos_survey_fnlfnl.printSchema()

<h4>Step 2: processing S&P Factset data

**Load S&P factset data**

SP500_factset = pd.read_excel(data_path + "S+P500_Factset.xlsx", engine="openpyxl")
SP500_factset = spark.createDataFrame(SP500_factset)

# sample print
SP500_factset.show()

# check schema
SP500_factset.printSchema

# check sector classification
SP500_factset.groupBy('RBICS Economy').count().sort('RBICS Economy').show()

# define the columns that need to be cleaned
fin_data_cols = ['Closing Price',
                 'Market Value',
                 'Sales',
                 'Unnamed: 7', ## what's this?
                 '1 Year Growth Total Sales',
                 'FE Sales Mean FY1 Roll',
                 'EBITDA Oper Income',
                 'FE Ebitda Mean FY1 Roll',
                 'Earns Per Share',
                 'FE Eps Mean FY1 Roll',
                 'FE Val Roe Mean FY1 Roll',
                 'Return on Avg Total Assets',
                 'FE Val Pe Mean FY1 Roll',
                 'Resrch and Develop Expense'
                 ]

# clean the columns as they have NaN rather than nulls
SP500_factset_preprocess = SP500_factset

for c in fin_data_cols:
    if c in SP500_factset_preprocess.columns:
        SP500_factset_preprocess = SP500_factset_preprocess.withColumn(c, F.when(F.isnan(F.col(c)), None).otherwise(F.col(c)))

# group by sector and get averages
sector_SNP = (SP500_factset_preprocess
                .withColumn("RBICS Economy", F.trim(F.col("RBICS Economy")))
                .groupBy('RBICS Economy')
                .agg(*[F.avg(F.col(c).cast("double")).alias(f"avg_{c.replace(' ', '_')}")
                       for c in fin_data_cols])
                .withColumnRenamed('RBICS Economy', 'Sector')
                .sort('Sector')
                )

# sample print
sector_SNP.show()

<h4>Step 3: get one table with AI usage data and financial data combined

## join the 2 tables:
joined_table = (btos_survey_fnlfnl
                .join(sector_SNP,["Sector"],"left")
                .sort('Sector','Question ID')
                )

# sample print
joined_table.show()

# one table with question 7:
# In the last two weeks, did this business use Artificial Intelligence (AI) in producing goods or services? 
# (Examples of AI: machine learning, natural language processing, virtual agents, voice recognition, etc.)

AI_usage_wth_fin_data = (joined_table
                         .filter(col('Question ID')==7))

AI_usage_wth_fin_data.show(5)


# one table with question 24:
# During the next six months, do you think this business will be using Artificial Intelligence (AI) in producing goods or services? 
# (Examples of AI: machine learning, natural language processing, virtual agents, voice recognition,

AI_acceptance_wth_fin_data = (joined_table
                              .filter(col('Question ID')==24))

AI_acceptance_wth_fin_data.show(5)

<h4>Step 4: save the tables

AI_usage_wth_fin_data_pd = AI_usage_wth_fin_data.toPandas()
AI_acceptance_wth_fin_data_pd = AI_acceptance_wth_fin_data.toPandas()

AI_usage_wth_fin_data_pd.to_excel(data_path + "Saved_Tables/01_AI_usage_wth_fin_data.xlsx", index=False)
AI_acceptance_wth_fin_data_pd.to_excel(data_path + "Saved_Tables/01_AI_acceptance_wth_fin_data.xlsx", index=False)
