==> Code/02_Data_Wrangling.ipynb <==
<h2>Data Cleaning and Wrangling

# Initializing spark session 
from pyspark.sql import SparkSession 
spark = SparkSession.builder.appName("Test").getOrCreate() 

# Import relevant packages 
import pandas as pd
from pyspark.sql import functions as F
from pyspark.sql.functions import * 
from pyspark.sql import DataFrame 
from pyspark.sql.types import StructType, StructField, StringType, DoubleType
import builtins

# define data path
data_path = "/Users/lanzia/Desktop/Study/FS 2025/01.1st Semester/Introduction to Data in Business Analytics/Project/Data/Saved_Tables/"

# read the data
AI_usage = pd.read_excel(data_path + "01_AI_usage_wth_fin_data.xlsx", engine="openpyxl")
AI_acceptance = pd.read_excel(data_path + "01_AI_acceptance_wth_fin_data.xlsx", engine="openpyxl")

AI_usage = spark.createDataFrame(AI_usage)
AI_acceptance = spark.createDataFrame(AI_acceptance)

# define the columns needed to be cleaned later
ranges = [(202319, 202326), (202401, 202426), (202501, 202518)]
numeric_fields = [StructField(str(i), DoubleType(), True) 
                  for start, end in ranges for i in range(start, end + 1)
                  ]
cols_to_clean = ['avg_' + str(i) for start, end in ranges for i in range(start, end + 1)]

# define financial data columns
fin_data_cols = ['avg_Closing_Price',
                 'avg_Market_Value',
                 'avg_Sales',
                 'avg_1_Year_Growth_Total_Sales',
                 'avg_FE_Sales_Mean_FY1_Roll',
                 'avg_EBITDA_Oper_Income',
                 'avg_FE_Ebitda_Mean_FY1_Roll',
                 'avg_Earns_Per_Share',
                 'avg_FE_Eps_Mean_FY1_Roll',
                 'avg_FE_Val_Roe_Mean_FY1_Roll',
                 'avg_Return_on_Avg_Total_Assets',
                 'avg_FE_Val_Pe_Mean_FY1_Roll',
                 'avg_Resrch_and_Develop_Expense'
                 ]

all_columns = cols_to_clean + fin_data_cols

# round the digits of usage rate
for c in all_columns:
    if c in AI_usage.columns:
        AI_usage = (AI_usage
                    .withColumn(c, F.when(F.isnan(F.col(c)), None).otherwise(F.col(c)))
                    .withColumn(c, round(c, 1))
                    )

AI_usage.show()

# round the digits of acceptance rate
for c in all_columns:
    if c in AI_acceptance.columns:
        AI_acceptance = (AI_acceptance
                    .withColumn(c, F.when(F.isnan(F.col(c)), None).otherwise(F.col(c)))
                    .withColumn(c, round(c, 1))
                    )

AI_acceptance.show()

# create average usage rate
AI_usage_wth_avg = (AI_usage
                    .withColumn('two_year_avg', round((builtins.sum((F.col(c) for c in cols_to_clean), F.lit(0.0)) / len(cols_to_clean)),1))
                    .select('Sector', 'two_year_avg', *cols_to_clean, *fin_data_cols)
                    .fillna(0.0)
                    )

AI_usage_wth_avg.show()

# create average acceptance rate
AI_acceptance_wth_avg = (AI_acceptance
                    .withColumn('two_year_avg', round((builtins.sum((F.col(c) for c in cols_to_clean), F.lit(0.0)) / len(cols_to_clean)),1))
                    .select('Sector', 'two_year_avg', *cols_to_clean, *fin_data_cols)
                    .fillna(0.0)
                    )

AI_acceptance_wth_avg.show()

# save the tables
AI_usage_wth_avg_pd = AI_usage_wth_avg.toPandas()
AI_acceptance_wth_avg_pd = AI_acceptance_wth_avg.toPandas()

# adjust for one missing value: Utilities sector did not disclose R&D
mean_value_usage = AI_usage_wth_avg_pd.loc[AI_usage_wth_avg_pd['avg_Resrch_and_Develop_Expense'] != 0, 'avg_Resrch_and_Develop_Expense'].mean()
AI_usage_wth_avg_pd.loc[(AI_usage_wth_avg_pd['Sector'] == 'Utilities') & (AI_usage_wth_avg_pd['avg_Resrch_and_Develop_Expense'] == 0.0),
                        'avg_Resrch_and_Develop_Expense'] = mean_value_usage

AI_usage_wth_avg_pd.loc[AI_usage_wth_avg_pd['Sector'] == 'Utilities', 'avg_Resrch_and_Develop_Expense'] = \
    AI_usage_wth_avg_pd.loc[AI_usage_wth_avg_pd['Sector'] == 'Utilities', 'avg_Resrch_and_Develop_Expense'].round(1)

AI_usage_wth_avg_pd

# adjust for one value
mean_value_acceptance = AI_acceptance_wth_avg_pd.loc[AI_acceptance_wth_avg_pd['avg_Resrch_and_Develop_Expense'] != 0, 'avg_Resrch_and_Develop_Expense'].mean()
AI_acceptance_wth_avg_pd.loc[(AI_acceptance_wth_avg_pd['Sector'] == 'Utilities') & (AI_acceptance_wth_avg_pd['avg_Resrch_and_Develop_Expense'] == 0.0),
                        'avg_Resrch_and_Develop_Expense'] = mean_value_acceptance

AI_acceptance_wth_avg_pd.loc[AI_acceptance_wth_avg_pd['Sector'] == 'Utilities', 'avg_Resrch_and_Develop_Expense'] = \
    AI_acceptance_wth_avg_pd.loc[AI_acceptance_wth_avg_pd['Sector'] == 'Utilities', 'avg_Resrch_and_Develop_Expense'].round(1)

AI_acceptance_wth_avg_pd

# save the updated tables:
AI_usage_wth_avg_pd.to_excel(data_path + "02_AI_usage_upd.xlsx", index=False)
AI_acceptance_wth_avg_pd.to_excel(data_path + "02_AI_acceptance_upd.xlsx", index=False)
